Moved WMI classes in WMI folder.
